add user authentication
Users are now able to create an account or sign in with an existing one,
and are restricted access to the site if not signed in. This is done
using NextAuth and (finally) connecting to a MongoDB database and Prisma
ORM, allowing for functionality that is more easily expanded upon. This
also includes the addition of auxiliary features such as extra component
styles/states, (temporary) sign out from the nav bar, and notifications,
as well as the removal of the old unused welcome page.